requirements.txt
flask
pandas
mysql-connector-python
pysftp
python-dotenv
